feat(gh-actions): widget prototype
